+++ src/js/components/Overview/Drawer/ManageChartsDrawer.js
